add DTOs

+++ src/main/java/org/example/practice/DTO/AccurateClientDTO.java
@@ -1,14 +1,12 @@
 package org.example.practice.DTO;
 
-import org.example.practice.entities.Client;
-
 public class AccurateClientDTO {
 
     public RiskDTO risk;
     public ClientDTO bestClient;
-    public int accuracyCount;
+    public float accuracyCount;
 
-    public AccurateClientDTO(RiskDTO risk, ClientDTO bestClient, int accuracyCount) {
+    public AccurateClientDTO(RiskDTO risk, ClientDTO bestClient, float accuracyCount) {
         this.risk = risk;
         this.bestClient = bestClient;
         this.accuracyCount = accuracyCount;
